Refactor form builder into separate components and fix module import paths
